+++ src/api/server/list_servers.rs
@@ -9,15 +9,15 @@ use typed_builder::TypedBuilder;
 #[serde(rename_all = "PascalCase")]
 pub struct ListServersRequest {
     #[serde(skip)]
-    pub count: isize,
+    pub count: i64,
     #[serde(skip)]
-    pub offset: isize,
+    pub offset: i64,
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "PascalCase")]
 pub struct ListServersResponse {
-    pub total_count: isize,
+    pub total_count: i64,
     pub servers: Vec<Server>,
 }
